Add CouchDB driver as config.data backend storage type
The user can now choose between 'disk' or 'couchdb' during cluster
initialization as config.data backend storage type.

If choosing 'couchdb', the config.data represantation changes, for
performance reasons, but only in the way it flushes the changes to the db.

Confd daemon isn't configured to work with CouchDB yet, so i recommend
to disable it during configure, (--disable-confd) flag.